Splitup PagerandoBinning into .h and .cpp file
