+++ lib/CodeGen/PagerandoBinning.cpp
@@ -20,6 +20,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "llvm/CodeGen/PagerandoBinning.h"
 #include "llvm/ADT/StringExtras.h"
 #include "llvm/CodeGen/MachineFunction.h"
 #include "llvm/CodeGen/MachineModuleInfo.h"
@@ -30,34 +31,6 @@
 using namespace llvm;
 
 #define DEBUG_TYPE "pagerando"
-
-namespace {
-class PagerandoBinning : public ModulePass {
-public:
-  static char ID;
-  explicit PagerandoBinning() : ModulePass(ID), BinCount(1) {
-    initializePagerandoBinningPass(*PassRegistry::getPassRegistry());
-  }
-
-  bool runOnModule(Module &M) override;
-
-  void getAnalysisUsage(AnalysisUsage &AU) const override {
-    AU.addRequired<MachineModuleInfo>();
-    AU.setPreservesAll();
-    ModulePass::getAnalysisUsage(AU);
-  }
-
-private:
-  static constexpr unsigned BinSize = 4096; // one page
-  static constexpr unsigned MinFnSize = 2;  // 'bx lr' on ARM thump
-
-  // Map <free space -> bin numbers>
-  std::multimap<unsigned, unsigned> Bins;
-  unsigned BinCount;
-
-  unsigned AssignToBin(const MachineFunction &MF);
-};
-} // end anonymous namespace
 
 char PagerandoBinning::ID = 0;
 INITIALIZE_PASS_BEGIN(PagerandoBinning, "pagerando-binning",
